+++ src/main/java/blackjack/domain/game/BlackJackReferee.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.game;
+
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Player;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
