+++ src/main/java/blackjack/domain/participant/Cards.java
@@ -1,4 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardSuit;
 
 import java.util.ArrayList;
 import java.util.List;
