+++ src/main/java/blackjack/domain/participant/Participants.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import java.util.ArrayList;
 import java.util.Collections;
