+++ src/test/java/blackjack/domain/DeckTest.java
@@ -2,6 +2,10 @@ package blackjack.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardSuit;
+import blackjack.domain.card.Deck;
 import blackjack.util.CardPickerGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
