+++ src/test/java/blackjack/domain/PlayerNameTest.java
@@ -3,6 +3,7 @@ package blackjack.domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
+import blackjack.domain.participant.ParticipantName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
